adicionadas correcoes de tipo de dx, dy, dz e dt que tinham tipo int ao inves de float. Correcao na politicao do raja que estavaa usando definindo threads ao inves de blocos

+++ cuda/cuda_driver.hpp
@@ -71,7 +71,7 @@ class CudaDriver : public Driver {
 
     private:
         int sx, sy, sz;
-        int dx, dy, dz, dt;
+        float dx, dy, dz, dt;
         int nx, ny, nz;
         int bord, absorb;
         host_data_container   host_data;
