+++ kokkos/kokkos_driver.hpp
@@ -75,7 +75,7 @@ class KokkosDriver : public Driver {
     private:
         Kokkos::ScopeGuard guard_;
         int sx, sy, sz;
-        int dx, dy, dz, dt;
+        float dx, dy, dz, dt;
         int nx, ny, nz;
         int bord, absorb;
         host_data_container   host_data;
